acornify emitDCEGraph [ci skip]
