+++ tools/js-optimizer.js
@@ -7969,229 +7969,6 @@ function isDynamicDynCall(node) {
          (node[0] === 'string' && node[1] === 'dynCall_');
 }
 
-//
-// Emit the DCE graph, to help optimize the combined JS+wasm.
-// This finds where JS depends on wasm, and where wasm depends
-// on JS, and prints that out.
-//
-// The analysis here is simplified, and not completely general. It
-// is enough to optimize the common case of JS library and runtime
-// functions involved in loops with wasm, but not more complicated
-// things like JS objects and sub-functions. Specifically we
-// analyze as follows:
-//
-//  * We consider (1) the toplevel scope, and (2) the scopes of toplevel defined
-//    functions (defun, not function; i.e., function X() {} where
-//    X can be called later, and not y = function Z() {} where Z is
-//    just a name for stack traces). We also consider the wasm, which
-//    we can see things going to and arriving from.
-//  * Anything used in a defun creates a link in the DCE graph, either
-//    to another defun, or the wasm.
-//  * Anything used in the toplevel scope is rooted, as it is code
-//    we assume will execute. The exceptions are
-//     * when we receive something from wasm; those are "free" and
-//       do not cause rooting. (They will become roots if they are
-//       exported, the metadce logic will handle that.)
-//     * when we send something to wasm; sending a defun causes a
-//       link in the DCE graph.
-//  * Anything not in the toplevel or not in a toplevel defun is
-//    considering rooted. We don't optimize those cases.
-//
-// Special handling:
-//
-//  * dynCall('vii', ..) are dynamic dynCalls, but we analyze them
-//    statically, to preserve the dynCall_vii etc. method they depend on.
-//    Truly dynamic dynCalls (not to a string constant) will not work,
-//    and require the user to export them.
-//  * Truly dynamic dynCalls are assumed to reach any dynCall_*.
-//
-// XXX this modifies the input AST. if you want to keep using it,
-//     that should be fixed. Currently the main use case here does
-//     not require that. TODO FIXME
-//
-function emitDCEGraph(ast) {
-  // First pass: find the wasm imports and exports, and the toplevel
-  // defuns, and save them on the side, removing them from the AST,
-  // which makes the second pass simpler.
-  //
-  // The imports that wasm receives look like this:
-  //
-  //  Module.asmLibraryArg = { "abort": abort, "assert": assert, [..] };
-  //
-  // The exports are trickier, as they have a different form whether or not
-  // async compilation is enabled. It can be either:
-  //
-  //  var _malloc = Module["_malloc"] = asm["_malloc"];
-  //
-  // or
-  //
-  //  var _malloc = Module["_malloc"] = (function() {
-  //   return Module["asm"]["_malloc"].apply(null, arguments);
-  //  });
-  //
-  var imports = [];
-  var defuns = [];
-  var dynCallNames = [];
-  var nameToGraphName = {};
-  var modulePropertyToGraphName = {};
-  var exportNameToGraphName = {}; // identical to asm['..'] nameToGraphName
-  var foundAsmLibraryArgAssign = false;
-  var graph = [];
-  traverse(ast, function(node, type) {
-    if (isAsmLibraryArgAssign(node)) {
-      var items = node[3][1];
-      items.forEach(function(item) {
-        imports.push(item[1][1]); // the name doesn't matter, only the value which is that actual thing we are importing
-      });
-      foundAsmLibraryArgAssign = true;
-      return emptyNode(); // ignore this in the second pass; this does not root
-    } else if (type === 'var') {
-      if (node[1] && node[1].length === 1) {
-        var item = node[1][0];
-        var name = item[0];
-        var value = item[1];
-        if (Array.isArray(value) && value[0] === 'assign') {
-          var assigned = value[2];
-          if (isModuleUse(assigned) && getModuleUseName(assigned) === name) {
-            // this is
-            //  var x = Module['x'] = ?
-            // which looks like a wasm export being received. confirm with the asm use
-            var found = 0;
-            var asmName;
-            traverse(value[3], function(node, type) {
-              if (isAsmUse(node)) {
-                found++;
-                asmName = getAsmUseName(node);
-              }
-            });
-            // in the wasm backend, the asm name may have one fewer "_" prefixed
-            if (found === 1) {
-              // this is indeed an export
-              // the asmName is what the wasm provides directly; the outside JS
-              // name may be slightly different (extra "_" in wasm backend)
-              var graphName = getGraphName(name, 'export');
-              nameToGraphName[name] = graphName;
-              modulePropertyToGraphName[name] = graphName;
-              exportNameToGraphName[asmName] = graphName;
-              if (/^dynCall_/.test(name)) {
-                dynCallNames.push(graphName);
-              }
-              return emptyNode(); // ignore this in the second pass; this does not root
-            }
-          }
-        }
-      }
-    } else if (type === 'defun') {
-      defuns.push(node);
-      var name = node[1];
-      nameToGraphName[name] = getGraphName(name, 'defun');
-      return emptyNode(); // ignore this in the second pass; we scan defuns separately
-    } else if (type === 'function') {
-      return null; // don't look inside
-    }
-  });
-  // must find the info we need
-  assert(foundAsmLibraryArgAssign, 'could not find the assigment to "asmLibraryArg". perhaps --pre-js or --post-js code moved it out of the global scope? (things like that should be done after emcc runs, as they do not need to be run through the optimizer which is the special thing about --pre-js/--post-js code)');
-  // Second pass: everything used in the toplevel scope is rooted;
-  // things used in defun scopes create links
-  function getGraphName(name, what) {
-    return 'emcc$' + what + '$' + name;
-  }
-  var infos = {}; // the graph name of the item => info for it
-  imports.forEach(function(import_) {
-    var name = getGraphName(import_, 'import');
-    var info = infos[name] = {
-      name: name,
-      import: ['env', import_],
-      reaches: {}
-    };
-    if (nameToGraphName.hasOwnProperty(import_)) {
-      info.reaches[nameToGraphName[import_]] = 1;
-    } // otherwise, it's a number, ignore
-  });
-  for (var e in exportNameToGraphName) {
-    var name = exportNameToGraphName[e];
-    infos[name] = {
-      name: name,
-      export: e,
-      reaches: {}
-    };
-  }
-  // a function that handles a node we visit, in either a defun or
-  // the toplevel scope (in which case the second param is not provided)
-  function visitNode(node, defunInfo) {
-    // TODO: scope awareness here. for now we just assume all uses are
-    //       from the top scope, which might create more uses than needed
-    var reached;
-    if (node[0] === 'name') {
-      var name = node[1];
-      if (nameToGraphName.hasOwnProperty(name)) {
-        reached = nameToGraphName[name];
-      }
-    } else if (isModuleUse(node)) {
-      var name = getModuleUseName(node);
-      if (modulePropertyToGraphName.hasOwnProperty(name)) {
-        reached = modulePropertyToGraphName[name];
-      }
-    } else if (isStaticDynCall(node)) {
-      reached = getGraphName(getStaticDynCallName(node), 'export');
-    } else if (isDynamicDynCall(node)) {
-      // this can reach *all* dynCall_* targets, we can't narrow it down
-      reached = dynCallNames;
-    } else if (isAsmUse(node)) {
-      // any remaining asm uses are always rooted in any case
-      var name = getAsmUseName(node);
-      if (exportNameToGraphName.hasOwnProperty(name)) {
-        infos[exportNameToGraphName[name]].root = true;
-      }
-      return;
-    }
-    if (reached) {
-      function addReach(reached) {
-        if (defunInfo) {
-          defunInfo.reaches[reached] = 1; // defun reaches it
-        } else {
-          infos[reached].root = true; // in global scope, root it
-        }
-      }
-      if (typeof reached === 'string') {
-        addReach(reached);
-      } else {
-        reached.forEach(addReach);
-      }
-    }
-  }
-  defuns.forEach(function(defun) {
-    var name = getGraphName(defun[1], 'defun');
-    var info = infos[name] = {
-      name: name,
-      reaches: {}
-    };
-    traverse(defun[3], function(node, type) {
-      visitNode(node, info);
-    });
-  });
-  traverse(ast, function(node, type) {
-    visitNode(node, null);
-  });
-  // Final work: print out the graph
-  // sort for determinism
-  function sortedNamesFromMap(map) {
-    var names = [];
-    for (var name in map) {
-      names.push(name);
-    }
-    names.sort();
-    return names;
-  }
-  sortedNamesFromMap(infos).forEach(function(name) {
-    var info = infos[name];
-    info.reaches = sortedNamesFromMap(info.reaches);
-    graph.push(info);
-  });
-  print(JSON.stringify(graph, null, ' '));
-}
-
 function removeFuncs(ast) {
   assert(ast[0] === 'toplevel');
   var keep = set(extraInfo.keep);
@@ -8239,7 +8016,6 @@ var passes = {
   findReachable: findReachable,
   dumpCallGraph: dumpCallGraph,
   asmLastOpts: asmLastOpts,
-  emitDCEGraph: emitDCEGraph,
   removeFuncs: removeFuncs,
   noop: function() {},
